rename batest to hudtest
