+++ district_test.go
@@ -4,13 +4,13 @@ import (
 	"encoding/json"
 	"testing"
 
-	"github.com/oysterriveroverdrive/hud/batest"
+	"github.com/oysterriveroverdrive/hud/hudtest"
 	"github.com/oysterriveroverdrive/hud/model"
 	"github.com/stretchr/testify/assert"
 )
 
 func TestDistricts(t *testing.T) {
-	blueAlliance, _ := batest.NewBlueAlliance()
+	blueAlliance, _ := hudtest.NewBlueAlliance()
 	defer blueAlliance.Close()
 
 	want := []*model.District{
